requirements.txt
flask>=3.0.0
gspread>=6.1.2
google-auth>=2.28.0
gunicorn>=21.2.0
